requirements.txt
tensorflow-gpu
bert-tensorflow
matplotlib
tqdm
pandas
underthesea
scikit-learn
numpy==1.16.4
h5py
keras-bert>=0.50.0
keras-gpt-2>=0.8.0
gensim>=3.5.0
seqeval==0.0.10
bert4keras==0.6.5
